Varios changes
