+++ src/main/java/com/github/skjolber/stcsv/CsvReaderFactory2.java
@@ -1,0 +1,15 @@
+package com.github.skjolber.stcsv;
+
+import java.io.Reader;
+
+/**
+ * 
+ * Dynamic CSV parser generator. Adapts the underlying implementation according 
+ * to the first (header) line.
+ */
+
+public interface CsvReaderFactory2<T, I> {
+
+	CsvReader<T> create(Reader reader, I intermediate) throws Exception;
+	
+}
