+++ src/main/java/com/github/skjolber/stcsv/column/tri/BooleanCsvColumnValueTriConsumer.java
@@ -1,0 +1,21 @@
+package com.github.skjolber.stcsv.column.tri;
+
+import com.github.skjolber.stcsv.column.bi.BooleanCsvColumnValueConsumer;
+
+public class BooleanCsvColumnValueTriConsumer<T, I> implements CsvColumnValueTriConsumer<T, I> {
+
+	protected final ObjBooleanTriConsumer<T, I> setter;
+		
+	public BooleanCsvColumnValueTriConsumer(ObjBooleanTriConsumer<T, I> setter) {
+		this.setter = setter;
+	}
+
+	@Override
+	public void consume(T object, I intermediate, char[] array, int start, int end) {
+		setter.accept(object, intermediate,
+				BooleanCsvColumnValueConsumer.parseBoolean(array, start, end)
+				);
+		
+	}
+
+}
