+++ src/main/java/com/github/skjolber/stcsv/column/tri/CsvColumnValueTriConsumer.java
@@ -1,17 +1,18 @@
-package com.github.skjolber.stcsv.column;
+package com.github.skjolber.stcsv.column.tri;
 
 @FunctionalInterface
-public interface CsvColumnValueConsumer<T> {
+public interface CsvColumnValueTriConsumer<T, I> {
 
 	/**
 	 * Generic interface for parsing values from a character array.
 	 * 
 	 * @param object target
+	 * @param intermediate intermediate helper / processor
 	 * @param array character array
 	 * @param start start index, inclusive
 	 * @param end end index, exclusive
 	 */
 	
-	void consume(T object, char[] array, int start, int end);
+	void consume(T object, I intermediate, char[] array, int start, int end);
 
 }
