+++ src/main/java/com/github/skjolber/stcsv/column/tri/DoubleCsvColumnValueTriConsumer.java
@@ -1,0 +1,16 @@
+package com.github.skjolber.stcsv.column.tri;
+
+import com.github.skjolber.stcsv.column.bi.DoubleCsvColumnValueConsumer;
+
+public class DoubleCsvColumnValueTriConsumer<T, I> implements CsvColumnValueTriConsumer<T, I> {
+
+	protected final ObjDoubleTriConsumer<T, I> setter;
+		
+	public DoubleCsvColumnValueTriConsumer(ObjDoubleTriConsumer<T, I> setter) {
+		this.setter = setter;
+	}
+
+	public void consume(T object, I intermediate, char[] array, int start, int end) {
+		setter.accept(object, intermediate, DoubleCsvColumnValueConsumer.parseDouble(array, start, end));
+	}
+}
