+++ src/main/java/com/github/skjolber/stcsv/column/tri/ObjBooleanTriConsumer.java
@@ -1,4 +1,4 @@
-package com.github.skjolber.stcsv.column;
+package com.github.skjolber.stcsv.column.tri;
 
 import java.util.function.BiConsumer;
 
@@ -16,13 +16,14 @@ import java.util.function.BiConsumer;
  *
  */
 @FunctionalInterface
-public interface ObjBooleanConsumer<T> {
+public interface ObjBooleanTriConsumer<T, I> {
 
     /**
      * Performs this operation on the given arguments.
      *
      * @param t the first input argument
+	 * @param intermediate intermediate helper / processor
      * @param value the second input argument
      */
-    void accept(T t, boolean value);
+    void accept(T t, I intermediate, boolean value);
 }
