+++ src/test/java/com/github/skjolber/stcsv/ClassicNoLineBreakQuotedColumn.java
@@ -6,7 +6,7 @@ import org.objectweb.asm.MethodVisitor;
 
 import com.github.skjolber.stcsv.AbstractColumn;
 import com.github.skjolber.stcsv.CsvException;
-import com.github.skjolber.stcsv.column.CsvColumnValueConsumer;
+import com.github.skjolber.stcsv.column.bi.CsvColumnValueConsumer;
 
 public class ClassicNoLineBreakQuotedColumn extends AbstractColumn {
 
@@ -374,14 +374,14 @@ public class ClassicNoLineBreakQuotedColumn extends AbstractColumn {
 
 	@Override
 	public void middle(MethodVisitor mv, String subClassInternalName, boolean inline) {
-		if(consumer == null) {
+		if(biConsumer == null) {
 			throw new IllegalArgumentException();
 		}
 	}
 
 	@Override
 	public void last(MethodVisitor mv, String subClassInternalName, boolean carriageReturn, boolean inline) {
-		if(consumer == null) {
+		if(biConsumer == null) {
 			throw new IllegalArgumentException();
 		}
